bismillah v1

+++ app/Models/Partner.php
@@ -16,7 +16,7 @@ class Partner extends Model
     protected static function booted()
     {
         static::deleting(function ($partner) {
-            if ($partner->logo) {
+            if ($partner->logo && $partner->logo !== '/default/extracurricular.svg') {
                 Storage::disk('public')->delete($partner->logo);
             }
         });
@@ -24,10 +24,10 @@ class Partner extends Model
         static::updating(function ($partner) {
             if ($partner->isDirty('logo')) {
                 $oldLogo = $partner->getOriginal('logo');
-                if ($oldLogo) {
+                if ($oldLogo && $partner->logo !== '/default/extracurricular.svg') {
                     Storage::disk('public')->delete($oldLogo);
                 }
             }
         });
-    }   
+    }
 }
